+++ src/app/shared/components/json-viewer/index.ts
@@ -1,0 +1,2 @@
+export * from './json-viewer.module';
+export * from './types/patch';
