+++ src/app/shared/components/json-viewer/types/item.ts
@@ -1,0 +1,5 @@
+import { Patch } from './patch';
+
+export interface Item extends Patch {
+    isPatched: boolean;
+}
