+++ src/app/shared/components/json-viewer/types/patch.ts
@@ -1,0 +1,6 @@
+export interface Patch {
+    path: string[];
+    key?: string;
+    value?: unknown;
+    tooltip?: string;
+}
